# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mysunriser Backend API
  version: 1.0.0
  description: OpenAPI spec generated from current controllers.
servers:
  - url: http://localhost:8080
paths:
  /api/health:
    get:
      summary: Health check
      description: Returns API and database health status.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/page:
    get:
      summary: List page items
      description: Returns a paged list of blog items. Anonymous and user requests only include Published posts; admin requests include all statuses.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/blog:
    post:
      summary: Create or update post (legacy admin-only)
      description: |
        Legacy upsert endpoint retained for compatibility.
        Requires an admin bearer token.
        This endpoint maps to backend insertOrUpdate behavior.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "200":
          description: Operation result text
          content:
            text/plain:
              schema:
                type: string
                example: Success!
        "400":
          description: Validation error or malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/blog/{slug}:
    get:
      summary: Get post by slug
      description: Anonymous and user requests can only read Published posts. Admin requests can read Draft and Published posts.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/posts/{slug}:
    put:
      summary: Update post as admin
      description: Updates editable detail fields for an existing post. Requires an admin bearer token.
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete post as admin
      description: Deletes an existing post and its post-media references. Requires an admin bearer token.
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/posts:
    post:
      summary: Create post as admin
      description: Creates a new post. Requires an admin bearer token and rejects duplicate slugs.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "200":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "400":
          description: Validation error or duplicate slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/media:
    post:
      summary: Upload media as admin
      description: Uploads an image or attachment. Requires an admin bearer token.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                accessLevel:
                  $ref: "#/components/schemas/MediaAccessLevel"
              required:
                - file
      responses:
        "200":
          description: Uploaded media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: List media as admin
      description: Returns a paged media library list. Requires an admin bearer token.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMediaListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/media/{id}:
    delete:
      summary: Delete unreferenced media as admin
      description: Deletes media only when no post references it. Requires an admin bearer token.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Deleted
        "400":
          description: Referenced media cannot be deleted or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/media/{id}/content:
    get:
      summary: Read media content
      description: |
        Returns media bytes inline. Admin can always read.
        Non-admin access requires the media to be referenced by a Published post.
        PUBLIC media can be read anonymously; AUTHENTICATED media requires any logged-in user.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Media bytes
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "401":
          description: Login required for authenticated media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/media/{id}/download:
    get:
      summary: Download media
      description: Same access rules as /api/media/{id}/content, but returns attachment disposition.
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Media bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Login required for authenticated media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/projects:
    get:
      summary: List published projects
      description: Returns a paged list of Published projects. README bodies are omitted from list items.
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
  /api/projects/{slug}:
    get:
      summary: Get project by slug
      description: Returns a published project and its cached or freshly fetched GitHub README Markdown.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/projects:
    get:
      summary: List projects as admin
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
    post:
      summary: Create project as admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
  /api/admin/projects/{slug}:
    put:
      summary: Update project as admin
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
    delete:
      summary: Delete project as admin
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthResponse:
      type: object
      properties:
        ok:
          type: boolean
        db:
          type: boolean
      required:
        - ok
        - db
    PageResponse:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PageItem"
      required:
        - page
        - pageSize
        - items
    PageItem:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        summary:
          type: string
        status:
          type: string
        publishAt:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      description: |
        Create payload for admin post creation and legacy /api/blog POST.
        Request field uses published_at, while read responses use publishAt.
      properties:
        slug:
          type: string
          maxLength: 128
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        title:
          type: string
          maxLength: 200
        content:
          type: string
        status:
          type: string
          maxLength: 20
          description: Post status such as draft or published.
        published_at:
          type: string
          format: date-time
          nullable: true
          description: Admin create stores null when omitted; legacy /api/blog POST defaults to current server time.
      required:
        - slug
        - title
        - content
        - status
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        content:
          type: string
        status:
          type: string
          maxLength: 20
        published_at:
          type: string
          format: date-time
          nullable: true
      required:
        - title
        - content
        - status
    PostResponse:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        status:
          type: string
        publishAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - content
        - status
        - publishAt
    ProjectStatus:
      type: string
      enum:
        - Draft
        - Published
    AdminProjectRequest:
      type: object
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{0,99}$"
        title:
          type: string
          maxLength: 120
        summary:
          type: string
          maxLength: 500
        status:
          $ref: "#/components/schemas/ProjectStatus"
        repoOwner:
          type: string
        repoName:
          type: string
        repoUrl:
          type: string
          maxLength: 500
          description: Optional; defaults to https://github.com/{repoOwner}/{repoName} when blank.
        sortOrder:
          type: integer
      required:
        - slug
        - title
        - summary
        - repoOwner
        - repoName
    ProjectListResponse:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProjectResponse"
      required:
        - page
        - pageSize
        - total
        - items
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        title:
          type: string
        summary:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        repoOwner:
          type: string
        repoName:
          type: string
        repoUrl:
          type: string
        sortOrder:
          type: integer
        readmeMarkdown:
          type: string
          description: Empty on list responses; populated on detail when available.
        readmeError:
          type: string
          description: Non-empty when GitHub README refresh failed.
        readmeCachedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - summary
        - status
        - repoOwner
        - repoName
        - repoUrl
        - sortOrder
        - readmeMarkdown
        - readmeError
    MediaAccessLevel:
      type: string
      enum:
        - PUBLIC
        - AUTHENTICATED
    MediaAssetType:
      type: string
      enum:
        - IMAGE
        - ATTACHMENT
    MediaAssetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        assetType:
          $ref: "#/components/schemas/MediaAssetType"
        originalFilename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        accessLevel:
          $ref: "#/components/schemas/MediaAccessLevel"
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        contentUrl:
          type: string
          example: /api/media/1/content
        downloadUrl:
          type: string
          example: /api/media/1/download
        markdown:
          type: string
          example: "![image.png](/api/media/1/content)"
      required:
        - id
        - assetType
        - originalFilename
        - mimeType
        - sizeBytes
        - accessLevel
        - uploadedBy
        - createdAt
        - contentUrl
        - downloadUrl
        - markdown
    AdminMediaListResponse:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaAssetResponse"
      required:
        - page
        - pageSize
        - total
        - items
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: Business error code.
        message:
          type: string
        timeStamp:
          type: integer
          format: int64
      required:
        - code
        - message
        - timeStamp